requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==21.2.0
beautifulsoup4==4.10.0
behave==1.2.6
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
driver==0.1.0
idna==3.2
iniconfig==1.1.1
lxml==4.6.3
packaging==21.0
parse-type==0.5.2
parse==1.19.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest-html==3.1.1
pytest~=7.2.1
requests~=2.28.2
self==2020.12.3
selenium~=4.7.2
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.7
webdriver-manager==3.4.2

configparser~=5.2.0
Appium-Python-Client~=2.8.1
openpyxl~=3.0.10